Compute totals in buildService and use BigInteger
Change ServiceWithProperties.total to BigInteger and update buildService to accept DashboardRequest and compute accurate totals per service type via dashboardDataRepository (instead of using the properties list size). Also pass dashboardRequest when building services, tidy formatting/whitespace, and refactor several stream/collector blocks and sorting logic for assessments, payments and appeals to improve readability.

+++ municipal-services/property-services/src/main/java/org/egov/pt/models/ServiceWithProperties.java
@@ -13,7 +13,7 @@ import lombok.Setter;
 @NoArgsConstructor
 @AllArgsConstructor
 public class ServiceWithProperties {
-	private Integer total;
+	private BigInteger total;
 	private String type = null;
 	private List<PropertyData> properties;
